third_party/WebKit/LayoutTests/http/tests/devtools/console/console-eval-exception-report-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Tests that evaluating an expression with an exception in the console provide correct exception information.

        function foo()
        {
            throw {a:239};
        }
        function boo()
        {
            foo();
        }
        boo();
VM:3 Uncaught {a: 239}
foo @ VM:3
boo @ VM:7
(anonymous) @ VM:9
call frame:foo at :2
call frame:boo at :6
call frame: at :8
